requirements.txt
pexels_api==1.0.1
python-dotenv==0.20.0
requests==2.26.0
